refactored help_requests table, updated model tests and help request model and repository
