+++ lib/models/help_request.py
@@ -1,16 +1,17 @@
 class HelpRequest:
 
-    def __init__(self, id, date, title, messsage, date_range, user_id, max_price):
+    def __init__(self, id, date, title, messsage, start_date, end_date, user_id, maxprice):
         self.id = id
         self.date = date
         self.title = title
         self.message = messsage
-        self.date_range = date_range
+        self.start_date = start_date
+        self.end_date = end_date
         self.user_id = user_id
-        self.max_price = max_price
+        self.maxprice = maxprice
 
     def __eq__(self, other):
         return self.__dict__ == other.__dict__
     
     def __repr__(self):
-        return f"HelpRequest({self.id}, {self.date}, {self.title}, {self.message}, {self.date_range}, {self.user_id}, {self.max_price})"
+        return f"HelpRequest({self.id}, {self.date}, {self.title}, {self.message}, {self.start_date}, {self.end_date}, {self.user_id}, {self.maxprice})"
